requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@84caae8038d0d3adc860957ccef05baeec2e2dd8
discord.py==1.5.1
idna==2.10
idna-ssl==1.1.0
jikanpy==4.2.2
multidict==4.7.6
pkg-resources==0.0.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.10
yarl==1.5.1
